requirements.txt
streamlit
langchain
langchain-groq
langchain-core
pydantic
python-dotenv
faiss-cpu
sentence-transformers
PyPDF2
groq
streamlit
numpy
tqdm
plotly